requirements.txt
anyio==3.7.1
APScheduler==3.6.3
asttokens==2.4.0
backcall==0.2.0
cachetools==4.2.2
certifi==2023.5.7
colorama==0.4.6
comm==0.1.4
debugpy==1.7.0
decorator==5.1.1
dnspython==2.4.2
executing==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
pymongo==4.5.0
python-dateutil==2.8.2
python-telegram-bot==20.4
pytz==2023.3
pywin32==306
pyzmq==25.1.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
tzdata==2023.3
tzlocal==5.0.1
wcwidth==0.2.6
